refactor: #668 notification View 코드 xml 안으로 옮기기
