+++ android/TodokTodok/app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.view.notification.NotificationActivity"
+            android:name=".presentation.xml.notification.NotificationActivity"
             android:exported="false" />
         <service
             android:name=".adapter.FirebaseAdapter"
