+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/xml/notification/NotificationUiState.kt
@@ -1,7 +1,7 @@
-package com.team.todoktodok.presentation.view.notification
+package com.team.todoktodok.presentation.xml.notification
 
 import com.team.domain.model.notification.Notification
-import com.team.todoktodok.presentation.view.notification.adapter.NotificationGroup
+import com.team.todoktodok.presentation.xml.notification.adapter.NotificationGroup
 
 data class NotificationUiState(
     val isLoading: Boolean = false,
