+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/xml/notification/adapter/NotificationViewHolder.kt
@@ -1,4 +1,4 @@
-package com.team.todoktodok.presentation.view.notification.adapter
+package com.team.todoktodok.presentation.xml.notification.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
